Migrated to Mac OS
Updated and added videos to a list view in the court_calendar_page. Added a time selector by day(All day, Morning/ Afternoon/ Evening)
Added an App icon image so it looks a bit better than the F, Also tried to style it but not going as thought. UI designer might be needed later in the app lifecyle.
Got the app to work on an iPhone emulator!
Next I am going to work on loading the videos and playing them when a user selects whichever video.