Added Video_Player_screen
Added the video player screen. Added the dependencies of the package as well.

Note: when adding this it created some form of MacOS folder were it's giving me an error because it's not finding the package. I read up a bit and it's nothing to worry about for now. Since it's only if we want to deploy to deskptop.